Now check for database existence in main at the start
This effectively removes all the database check I put all over the places
inside subfunctions, makes the code a bit cleaner. However, this comes
with a downside that now we to check for subcommand `init` separately
first. If you are reading this and have a better way to approach this
then please send a pull request.
Add a comment for Future features and fix a small grammar typo.